fix(user) : weight height 타입 변경

+++ src/main/java/com/dekk/user/application/command/UserOnboardingCommand.java
@@ -2,5 +2,10 @@ package com.dekk.user.application.command;
 
 import com.dekk.user.domain.model.enums.Gender;
 
-public record UserOnboardingCommand(String nickname, Double height, Double weight, Gender gender) {
+public record UserOnboardingCommand(
+        String nickname,
+        Integer height,
+        Integer weight,
+        Gender gender
+) {
 }
